Add login/logout views, is Ownerpermission, is_owner field to ProfileSerializer

+++ profiles/views.py
@@ -1,9 +1,11 @@
 from rest_framework.views import APIView
 from rest_framework import status
 from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated
+from django.http import Http404
 from .models import Profile
 from .serializers import ProfileSerializer
-from django.http import Http404
+from .permissions import IsOwnerOrReadOnly
 
 
 class ProfileList(APIView):
@@ -12,7 +14,8 @@ class ProfileList(APIView):
     """
     def get(self, request):
         profiles = Profile.objects.all()
-        serializer = ProfileSerializer(profiles, many=True)
+        serializer = ProfileSerializer(
+            profiles, many=True, context={'request': request})
         return Response(serializer.data)
 
 
@@ -20,28 +23,33 @@ class ProfileDetail(APIView):
     """
     View to retrieve, update, or delete a profile instance.
     """
+    permission_classes = [IsOwnerOrReadOnly]
     serializer_class = ProfileSerializer
+
     def get_object(self, pk):
         try:
             return Profile.objects.get(pk=pk)
         except Profile.DoesNotExist:
             raise Http404
 
-    def get(self, request, pk):
+    def get(self, request, pk, format=None):
         profile = self.get_object(pk)
-        serializer = ProfileSerializer(profile)
+        self.check_object_permissions(self.request, profile)
+        serializer = ProfileSerializer(profile, context={'request': request})
         return Response(serializer.data)
 
-    def put(self, request, pk):
+    def put(self, request, pk, format=None):
         profile = self.get_object(pk)
-        serializer = ProfileSerializer(profile, data=request.data)
+        self.check_object_permissions(request, profile)
+        serializer = ProfileSerializer(
+        profile, data=request.data, context={'request': request})
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request, pk):
+    def delete(self, request, pk, format=None):
         profile = self.get_object(pk)
+        self.check_object_permissions(self.request, profile)
         profile.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-
